Move deprecated cards in marisa into own package
